feat: refactor result rows to split out logic
